+++ src/main/java/bridge/config/DependencyInjector.java
@@ -1,6 +1,8 @@
-package bridge;
+package bridge.config;
 
 
+import bridge.InputView;
+import bridge.OutputView;
 import bridge.controller.Controller;
 import bridge.controller.RetryInputUtil;
 import bridge.domain.BridgeMaker;
